+++ examples/game/contexts/SceneContext.ts
@@ -1,4 +1,4 @@
-import {AbstractContext} from "../../../src";
+import {AbstractContext, inject, optional} from "../../../src";
 import {TimerLevelLoader} from "../adapters/TimerLevelLoader";
 import {Fire} from "../commands/Fire";
 import {Load} from "../commands/Load";
@@ -6,10 +6,17 @@ import {Tick} from "../commands/Tick";
 import {LEVEL_LOADER, SCENE_MODEL, SCENE_MODEL_IMMUTABLE, SCENE_NAME, SceneModelImmutable} from "../contracts";
 import {FIRE, LOAD, TICK} from "../messages";
 import {SceneModel} from "../models/SceneModel";
+import {SceneMediator} from "../mediators/SceneMediator";
+import {IGameViewFactory} from "../views/IGameViewFactory";
+import {VIEW_FACTORY} from "../views/ViewTokens";
 
 export class SceneContext extends AbstractContext
 {
     public model!: SceneModelImmutable;
+
+    @inject(VIEW_FACTORY)
+    @optional()
+    private viewFactory?: IGameViewFactory;
 
     protected override init(): void
     {
@@ -28,5 +35,11 @@ export class SceneContext extends AbstractContext
         this.map(FIRE, Fire);
         this.map(TICK, Tick);
         this.map(LOAD, Load, {concurrency: "latest"});
+
+        if (this.viewFactory)
+        {
+            this.provide(VIEW_FACTORY, this.viewFactory, ["mediator"]);
+            this.createMediator(SceneMediator, "scene");
+        }
     }
 }