+++ examples/game/views/ViewTokens.ts
@@ -1,3 +1,5 @@
-import {ServiceToken} from "../../../src";
+import {IMessageDispatcherImmutable, ServiceToken} from "../../../src";
+import type {IGameViewFactory} from "./IGameViewFactory";
 
-export const MOUNT = new ServiceToken<HTMLElement>("ViewMount");
+export const VIEW_FACTORY = new ServiceToken<IGameViewFactory>("GameViewFactory");
+export const GAME_EVENTS = new ServiceToken<IMessageDispatcherImmutable>("GameEvents");